Try in vain to get the slide transitions to animate smoothly; summarily fail.
If you want to try and salvage this burning wreck, when you move to a new
slide, it should fluidly scroll to the next slide in like a fifth of a second.
Bonus points if there's an easing algorithm. It's cool to let closures handle
multiple presses, as it'll make the animation ZOOM, which is rad as hell.

Honestly, I'm too fucking tired to try and make this portion work, especially
seeing as how I think it might require an entire rewrite. Oh yeah, and when I
see transitions like this in proper Powerpoint, they piss me off. ~oh well~